requirements.txt
opencv-python
numpy
Pillow
tqdm
moderngl
pyglet
click